[SC-170] Add flag to enable/disable OpenGL support for disp library. BabyMEG's acquisition PC does not have a strong GPU. Hence we need to stick to the CPU version. However, flag is activated by default. Also include reset flag in start_writing_raw again.

+++ applications/mne_scan/plugins/babymeg/babymeginfo.cpp
@@ -293,7 +293,7 @@ void BabyMEGInfo::MGH_LM_Parse_Para(QByteArray cmdstr)
         //qDebug()<<t_ch.ch_name;
         t_ch.scanNo = i;
         t_ch.logNo = i+1;
-        t_ch.cal = lm_ch_calicoef.at(i).toDouble();
+        t_ch.cal = lm_ch_calicoef.at(i).toFloat();
         t_ch.range =1.0f/lm_ch_gain.at(i).toFloat();//1; // set gain
 
         //qDebug()<<i<<":="<<t_ch.ch_name<<","<<t_ch.range<<","<<t_ch.cal;
